requirements.txt
aiosmtplib==5.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
Django==6.0.3
django-bolt==0.6.5
django-redis==6.0.0
dotenv==0.9.9
ecdsa==0.19.1
esprima==4.0.1
filelock==3.25.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
injector==0.24.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.20.0
msgspec-ext==0.5.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
phonenumbers==9.0.25
pillow==12.1.1
prompt_toolkit==3.0.52
pyasn1==0.6.2
pycparser==3.0
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
redis==7.2.1
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
tree-sitter==0.25.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0